Structure preparation for billing test extension
