+++ testsuite/tests/ui/billing/stripe/conftest.py
@@ -1,4 +1,4 @@
-"""Billing tests for Stripe payment gateway"""
+"""Conftest for Stripe gateway billing tests"""
 import pytest
 
 from testsuite.ui.objects import CreditCard
@@ -30,26 +30,3 @@ def setup_card(account, custom_devel_login, billing_address, navigator):
         return cc_view
 
     return _setup
-
-
-@pytest.mark.parametrize("cc_number, verify_3ds", [
-    ("4000002500003155", True),
-    ("4242424242424242", False),
-])
-# pylint: disable=too-many-arguments
-def test_stripe(setup_card, cc_number, verify_3ds, stripe, create_ui_invoice, create_api_invoice):
-    """
-    Tests basic billing scenario for Stripe gateway:
-        - Add CC details for an account
-        - Complete 3DS challenge if it's supported
-        - Trigger billing via UI
-        - Trigger billing via API
-    """
-    setup_card(cc_number, verify_3ds)
-
-    invoice = create_ui_invoice()
-    stripe.assert_payment(invoice)
-
-    invoice = create_api_invoice()
-    charged = invoice.charge()
-    stripe.assert_payment(charged)
